requirements.txt
fastapi
uvicorn[standard]
pydantic
twilio
paho-mqtt
opencv-python
ultralytics
requests
pyjwt
passlib[bcrypt]
python-multipart